- Rename SimulationLabel component to MeasurementValueTable - Move SimulationLabel's markup creation from its container into its own component
