+++ frontend/src/app/simulation/measurement-value-table/MeasurementValueTable.common.scss
@@ -1,25 +1,31 @@
-.simulation-label {
+.measurement-value-table {
     position: absolute;
     top: 0;
     left: 0;
     border-radius: 2px;
     white-space: pre-wrap;
     font-family: 'Roboto', sans-serif;
-    font-size: 0.5em;
+    font-size: 0.7em;
     max-width: 100vw;
+    pointer-events: none;
 
     th,
     td {
         padding: 2px 5px;
     }
 
+    thead,
+    tr {
+        background-color: transparent !important;
+    }
+
 }
 
-.simulation-label__heading {
+.measurement-value-table__heading {
     font-weight: bold;
     text-align: center;
 }
 
-.simulation-label__content {
+.measurement-value-table__content {
     padding: 2px;
 }
