+++ frontend/src/app/simulation/measurement-value-table/MeasurementValueTable.dark.scss
@@ -1,6 +1,6 @@
-@import 'SimulationLabel.common';
+@import 'MeasurementValueTable.common';
 
-.simulation-label {
+.measurement-value-table {
     box-shadow: 0 1px 3px rgba(0, 0, 0, 0.3), 0 -0.5px 1px rgba(0, 0, 0, 0.2);
     background-color: #555;
     color: #fff;
