+++ frontend/src/app/simulation/measurement-value-table/index.ts
@@ -1,0 +1,1 @@
+export * from './MeasurementValueTableContainer';
