--- frontend/src/app/simulation/simulation-labels/index.ts
@@ -1,1 +1,0 @@
-export * from './SimulationLabelsContainer';
